+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 #### ScriptingEngine
 
 * Added new ScEng key KEY_ADJUST_RETRIEVED_INTEGER which allows the script to tweak the retrieved integer before applying it.
+* Added KEY_DIVIDER for per calculations, to allow to check "per X cards" and only count full sets.
 ## 1.14
 
 ### Important for Upgrades